+++ main.py
@@ -62,7 +62,7 @@ import re
 
 
 # 注册插件，指定插件名、作者、描述和版本号
-@register("mrfzccl", "Lishining", "你知道的,我一直是明日方舟高手", "1.2.40")
+@register("mrfzccl", "Lishining", "你知道的,我一直是明日方舟高手", "1.2.42")
 class Mrfzccl(Star):
     _question_candidate_names: np.ndarray
     _question_candidate_urls: List[List[str]]
@@ -274,6 +274,7 @@ class Mrfzccl(Star):
             t2i_enabled=t2i_config.enabled,
             t2i_endpoint=t2i_config.endpoint,
             t2i_max_concurrent=t2i_config.max_concurrent,
+            html_render_func=self.html_render,
         )
         logger.info(f"[Mrfzccl] 渲染主题: {renderer_theme}")